Split out chaining unit tests
